Update encryption for credentials file
This is because we had to change the master encryption key, as it was
previously committed into Git. After updating the encryption key our
production deployment was failing because the new key can't decrypt the
old file.

As such, I've regenerated the credentials file which is now encrypted
with the new master key.